Abstract user auth to separate class
